moved some test files
