+++ test/DisciplinesSearchTestRouteTest.js
@@ -1,5 +1,5 @@
 var should = require('should');
-var discipline = require('../lib/discipline_test.js');
+var discipline = require('./discipline_test.js');
 var server = require('../lib/routes.js');
 var Client = require('mariasql');
