+++ test/JournalSearchTestRouteTest.js
@@ -1,5 +1,5 @@
 var should = require('should');
-var journal = require('../lib/journal_test.js');
+var journal = require('./journal_test.js');
 var server = require('../lib/routes.js');
 var dbscheme = require('../lib/db.js');
 var Client = require('mariasql');
